2019 Day 9: Sensor Boost
Pretty simple day, but with a lot of changes everywhere. Adding a new parameter mode and instruction was trivial.

Making it possible to address memory outside the original program was easy as
well, just change List<Int> to Map<Int, Int> and modify the overloaded operators
for array access.

Changing all Ints to Longs was not difficult but required a lot of changes everywhere.

Next time there is a similar challenge I need to remember to start out with Longs and Maps from the beginning instead of Ints and Arrays.

+++ src/main/aoc2019/Day2.kt
@@ -3,21 +3,21 @@ package aoc2019
 
 class Day2(input: List<String>) {
 
-    private val parsedInput = input.map { it.toInt() }.toMutableList()
+    private val parsedInput = input.map { it.toLong() }.toMutableList()
 
-    fun solvePart1(): Int {
+    fun solvePart1(): Long {
         val computer = Intcode(parsedInput)
-        computer.reinitialize(listOf(Pair(1, 12), Pair(2, 2)))
+        computer.reinitialize(listOf(Pair(1L, 12L), Pair(2L, 2L)))
         computer.run()
         return computer.dumpMemory()[0]
     }
 
     fun solvePart2(): Int {
-        val desired = 19690720
+        val desired = 19690720L
         val computer = Intcode(parsedInput)
         for (noun in 0 until 99) {
             for (verb in 0 until 99) {
-                computer.reinitialize(listOf(Pair(1, noun), Pair(2, verb)))
+                computer.reinitialize(listOf(Pair(1L, noun.toLong()), Pair(2L, verb.toLong())))
                 computer.run()
                 if (computer.dumpMemory()[0] == desired) {
                     return 100 * noun + verb
